work in progress on assembly synchronization
