+++ packages/core/assemblyManager/index.ts
@@ -1,0 +1,2 @@
+export { default } from './assemblyManager'
+export { default as assemblyConfigSchemas } from './assemblyConfigSchemas'
